Uprava zobrazeni spisu - nezavislost na hlavnim rodici (puvodni chybna implementace spisoveho planu, ktera u spisu nema byt)
git-svn-id: svn://svn.blue-point.cz/spisovka/branches/s3_brno@372 f57ce333-4fc4-42a6-9b5e-014a8da81a87

+++ app/models/DokumentSubjekt.php
@@ -49,11 +49,11 @@ class DokumentSubjekt extends BaseModel
     public function pripojit($dokument_id, $subjekt_id, $typ = 'AO') {
 
         $row = array();
-        $row['dokument_id'] = $dokument_id;
-        $row['subjekt_id'] = $subjekt_id;
+        $row['dokument_id'] = (int)$dokument_id;
+        $row['subjekt_id'] = (int)$subjekt_id;
         $row['typ'] = $typ;
         $row['date_added'] = new DateTime();
-        $row['user_id'] = Environment::getUser()->getIdentity()->id;
+        $row['user_id'] = (int)Environment::getUser()->getIdentity()->id;
 
         return $this->insert($row);
